svm: consolidate JNI classes as we only support a single JDK version now
